port IndentingLexerDedentAnchorTest to zio-test
